Update paths for new locations
Unfortunately, twisted.filepath doesn't seem to understand the `~` in linux
pathnames, so using a path relative to moving home directory is out. Thus the
organizer service will need to be updated as directories move. Perhaps it would
be good to have this as a command line option for the future.