Handle create in subdirs
This is necessary if we use '--inplace' with rsync, as the files will be
created in order and may not exist in time. There is a minor possibility of a
race condition (of course, as is always the case with recursive inotify). If
this becomes a real problem, we can check every time a file is added to a
directory (e.g. when a .NFO file is transferred).

Additionally, this cleans up some of the logging